# openapi.yaml
openapi: 3.1.0
info:
  title: Galaxy of Consequence RPG Backend API
  description: A comprehensive Star Wars RPG backend with NVIDIA Nemotron integration, persistent world simulation, reactive NPCs, and immersive Force mechanics following the Final Rule.
  version: 1.0.0
  contact:
    name: Galaxy of Consequence API
    email: api@galaxyofconsequence.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.galaxyofconsequence.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /generate_quest:
    post:
      tags:
        - Quest
      summary: Generate procedural quest
      description: Generate quest based on player state, faction relations, and Force alignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
              properties:
                user:
                  type: string
                session_id:
                  type: string
                difficulty:
                  type: string
                  enum: [easy, medium, hard, extreme]
                  default: medium
                quest_type:
                  type: string
                  enum: [combat, social, exploration, force, faction]
                location:
                  type: string
      responses:
        '200':
          description: Quest generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  quest:
                    $ref: '#/components/schemas/Quest'
  /accept_quest:
    post:
      tags:
        - Quest
      summary: Accept available quest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quest_id
                - user
              properties:
                quest_id:
                  type: integer
                user:
                  type: string
      responses:
        '200':
          description: Quest accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  quest_id:
                    type: integer
                  status:
                    type: string
                  title:
                    type: string
  /complete_quest:
    post:
      tags:
        - Quest
      summary: Complete active quest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quest_id
                - user
              properties:
                quest_id:
                  type: integer
                user:
                  type: string
                completion_method:
                  type: string
                  default: standard
                player_choices:
                  type: array
                  items:
                    type: string
                session_id:
                  type: string
      responses:
        '200':
          description: Quest completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  quest_id:
                    type: integer
                  title:
                    type: string
                  completion_result:
                    type: object
                  rewards_granted:
                    type: array
                  faction_changes:
                    type: object
                  force_alignment_change:
                    type: integer
                  consequences:
                    type: array

  # Session Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use 'Abracadabra' as the Bearer token
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /authenticate endpoint
  schemas:
...
    Quest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        giver:
          type: string
        rewards:
          type: object
        requirements:
          type: array
        difficulty:
          type: integer
        faction_impact:
          type: object
        force_impact:
          type: integer
        generation_reason:
          type: string
        status:
          type: string
...